requirements.txt
bidict==0.23.1
blinker==1.8.2
cachelib==0.9.0
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cvzone==1.6.1
cycler==0.12.1
dill==0.3.8
filelock==3.15.3
Flask==3.0.3
Flask-Caching==2.3.0
Flask-SocketIO==5.3.6
fonttools==4.53.0
fsspec==2024.6.0
h11==0.14.0
idna==3.7
intel-openmp==2021.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.9.0
mkl==2021.4.0
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
opencv-python==4.10.0.84
packaging==24.1
pandas==2.2.2
pillow==10.3.0
psutil==6.0.0
py-cpuinfo==9.0.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-engineio==4.9.1
python-socketio==5.11.2
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
scipy==1.13.1
seaborn==0.13.2
simple-websocket==1.0.0
six==1.16.0
sympy==1.12.1
tbb==2021.13.0
torch==2.3.1
torchvision==0.18.1
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
ultralytics==8.2.38
ultralytics-thop==2.0.0
urllib3==2.2.2
Werkzeug==3.0.3
wsproto==1.2.0
